refactor(temporal): move deepEqual and constants to utils.ts
types.ts stays pure types. deepEqual, retryPolicy, and threshold
constants move to utils.ts where they belong.

Committed-By-Agent: claude
