+++ apps/service/src/temporal/workflows.ts
@@ -9,8 +9,8 @@ import {
 } from '@temporalio/workflow'
 
 import type { SyncActivities } from './activities.js'
-import { deepEqual, CONTINUE_AS_NEW_THRESHOLD, EVENT_BATCH_SIZE, retryPolicy } from './types.js'
 import type { WorkflowStatus } from './types.js'
+import { deepEqual, CONTINUE_AS_NEW_THRESHOLD, EVENT_BATCH_SIZE, retryPolicy } from './utils.js'
 
 // Setup/teardown: 2m with retry
 const { setup, teardown } = proxyActivities<SyncActivities>({
